+++ lib/applications/routes.js
@@ -31,7 +31,12 @@ const all = (req, res, next) => {
         // add itentifier to each row
         const rows = utils.routes.identifier(rws, req.applications.merged);
         const applications = rows.map(rw => {
-          return appsutils.endpoints(rw);
+          let row = rw;
+          row.client = {
+            id: row['client-id'],
+            secret: row['client-secret'],
+          };
+          return row;
         });
 
         res.render('applications/all', {
@@ -149,7 +154,11 @@ const one = (req, res, next) => {
         reject(err);
       }
 
-      app = appsutils.endpoints(rows[0]);
+      app = rows[0];
+      app.client = {
+        id: app['client-id'],
+        secret: app['client-secret'],
+      };
 
       // add session data for this application
       _.set(req.session, 'form.applications.edit', {
